// src/sdk/types.ts
export type Maybe<T> = T | null
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K]
}
export type MakeOptional<T, K extends keyof T> = Omit<T, K> &
  { [SubKey in K]?: Maybe<T[SubKey]> }
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> &
  { [SubKey in K]: Maybe<T[SubKey]> }
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string
  String: string
  Boolean: boolean
  Int: number
  Float: number
  citext: string
  jsonb: { [key: string]: any } | null
  timestamptz: Date
  uuid: string
}
// ...
export type News = {
  __typename?: 'news'
  /** Содержание новости */
  content: Scalars['String']
  created_at: Scalars['timestamptz']
  id: Scalars['uuid']
  /** Заголовок новости */
  title: Scalars['String']
  updated_at: Scalars['timestamptz']
}
// ...
export type News_Set_Input = {
  content?: Maybe<Scalars['String']>
  created_at?: Maybe<Scalars['timestamptz']>
  id?: Maybe<Scalars['uuid']>
  title?: Maybe<Scalars['String']>
  updated_at?: Maybe<Scalars['timestamptz']>
}
// ...
export type Users = {
  __typename?: 'users'
  /** An object relationship */
  account?: Maybe<Auth_Accounts>
  avatar_url?: Maybe<Scalars['String']>
  created_at: Scalars['timestamptz']
  display_name?: Maybe<Scalars['String']>
  id: Scalars['uuid']
  phone_number?: Maybe<Scalars['String']>
  updated_at: Scalars['timestamptz']
}
// ...
export type NewsFragment = { __typename: 'news' } & Pick<
  News,
  'id' | 'title' | 'content'
>

export type UserFragment = { __typename: 'users' } & Pick<
  Users,
  'id' | 'display_name' | 'avatar_url' | 'phone_number'
> & { account?: Maybe<{ __typename?: 'auth_accounts' } & UserAccountFragment> }

export type UserAccountFragment = { __typename: 'auth_accounts' } & Pick<
  Auth_Accounts,
  'id' | 'email' | 'user_id'
>

export type CurrentUserFragment = { __typename?: 'users' } & UserFragment

export type CreateNewsPostMutationVariables = Exact<{
  title: Scalars['String']
  content: Scalars['String']
}>

export type CreateNewsPostMutation = { __typename?: 'mutation_root' } & {
  news?: Maybe<{ __typename?: 'news' } & NewsFragment>
}
// ...
export type UpdateNewsPostMutationVariables = Exact<{
  postId: Scalars['uuid']
  newsPost: News_Set_Input
}>

export type UpdateNewsPostMutation = { __typename?: 'mutation_root' } & {
  news_post?: Maybe<{ __typename?: 'news' } & NewsFragment>
}

export type GetCurrentUserQueryVariables = Exact<{
  userId: Scalars['uuid']
}>

export type GetCurrentUserQuery = { __typename?: 'query_root' } & {
  user?: Maybe<{ __typename?: 'users' } & CurrentUserFragment>
}

export type GetNewsQueryVariables = Exact<{ [key: string]: never }>

export type GetNewsQuery = { __typename?: 'query_root' } & {
  news: Array<{ __typename?: 'news' } & NewsFragment>
}

export type GetNewsPostQueryVariables = Exact<{
  postId: Scalars['uuid']
}>

export type GetNewsPostQuery = { __typename?: 'query_root' } & {
  news_post?: Maybe<{ __typename?: 'news' } & NewsFragment>
}